requirements.txt
fastapi
uvicorn[standard]
passlib[bcrypt]
pyjwt
httpx